+++ tools/tests/py2fgen/test_export.py
@@ -6,41 +6,5 @@
 # Please, refer to the LICENSE file in the root directory.
 # SPDX-License-Identifier: BSD-3-Clause
 
-import cffi
-import numpy as np
 
-from icon4py.tools import py2fgen
-from icon4py.tools.py2fgen import _export, test_utils
-
-
-def test_default_mapping_hook_array():
-    ffi = cffi.FFI()
-
-    array_ptr = ffi.new("int[10]")
-
-    array_mapper = _export.default_mapping(
-        None,
-        py2fgen.ArrayParamDescriptor(
-            rank=1, dtype=py2fgen.INT32, device=py2fgen.DeviceType.HOST, is_optional=False
-        ),
-    )
-    result = array_mapper(
-        test_utils.array_descriptor(ptr=array_ptr, shape=(10,), on_gpu=False, is_optional=False),
-        ffi=ffi,
-    )
-
-    assert isinstance(result, np.ndarray)
-
-
-_ffi_not_needed = None
-
-
-def test_default_mapping_hook_bool():
-    bool_mapper = _export.default_mapping(None, py2fgen.ScalarParamDescriptor(dtype=py2fgen.BOOL))
-
-    true_result = bool_mapper(-1, ffi=_ffi_not_needed)
-    assert isinstance(true_result, bool)
-    assert true_result
-    false_result = bool_mapper(0, ffi=_ffi_not_needed)
-    assert isinstance(false_result, bool)
-    assert not false_result
+# TODO
