+++ tools/tests/py2fgen/test_test_utils.py
@@ -44,13 +44,13 @@ def test_array_as_array_descriptor_lifetime(ffi):
     assert np.array_equal(arr, np.arange(100000, dtype=np.int32))
 
 
-def test_array_as_array_descriptor_do_not_keepalive(ffi):
-    # this test relies on overwriting the memory of the original array which went out of scope
-    # if flaky, we should remove it
-    result = array_to_array_descriptor(np.arange(100000, dtype=np.int32), ffi=ffi, keep_alive=False)
+# def test_array_as_array_descriptor_do_not_keepalive(ffi):
+#     # this test relies on overwriting the memory of the original array which went out of scope
+#     # if flaky, we should remove it
+#     result = array_to_array_descriptor(np.arange(100000, dtype=np.int32), ffi=ffi, keep_alive=False)
 
-    gc.collect()  # Force garbage collection
-    _dummy = np.arange(100000, dtype=np.int32) + 1  # let's hope we get the same memory
+#     gc.collect()  # Force garbage collection
+#     _dummy = np.arange(100000, dtype=np.int32) + 1  # let's hope we get the same memory
 
-    arr = utils.as_array(ffi, result, from_np_dtype(np.int32))
-    assert not np.array_equal(arr, np.arange(100000, dtype=np.int32))
+#     arr = utils.as_array(ffi, result, from_np_dtype(np.int32))
+#     assert not np.array_equal(arr, np.arange(100000, dtype=np.int32))
